+++ database/migrations/2022_04_25_205810_create_trip_photos_table.php
@@ -15,7 +15,8 @@ return new class extends Migration
     {
         Schema::create('trip_photos', function (Blueprint $table) {
             $table->id();
-            $table->integer('trip_id');
+//            $table->foreignId('trip_id')->references('id')->on('trips')->cascadeOnUpdate()->cascadeOnDelete();
+            $table->unsignedBigInteger('trip_id');
             $table->string('path');
             $table->timestamps();
         });
