+++ database/migrations/2022_04_25_211557_create_passengers_table.php
@@ -15,7 +15,9 @@ return new class extends Migration
     {
         Schema::create('passengers', function (Blueprint $table) {
             $table->id();
-            $table->integer('customer_trip_id');
+//            $table->foreignId('customer_trip_id')->references('id')->on('users')->cascadeOnDelete()->cascadeOnUpdate();
+            $table->unsignedBigInteger('customer_trip_id');
+
             $table->string('passenger_name');
             $table->timestamps();
         });
